finished classes and added qc out logic
New changes included finishing all the qc checks for data in both
supervised and unsupervised. There are still some weird logs being sent
to the qc file for missing times that need to be figured out.

Added logic to clean and save the qc master json as a csv for easy
parsing by the lab.